requirements.txt
defusedxml
dlib
face-recognition
ImageHash
numpy
olefile
Pillow
pillow_heif
psycopg[binary,pool]
rawpy
PyExifTool
pydantic
pgvector
geopy